refactor: extract arg definitions into new module
