+++ tests/test_main_kr.py
@@ -4,17 +4,18 @@ from pathlib import Path
 from tempfile import TemporaryDirectory
 
 import pytest
-from _pytest.capture import CaptureFixture
-from _pytest.monkeypatch import MonkeyPatch
 from freezegun import freeze_time
 
-from datamodel_code_generator import __main__, inferred_message
+from datamodel_code_generator import inferred_message
 from datamodel_code_generator.__main__ import Exit, main
 
 try:
     from pytest import TempdirFactory
 except ImportError:
     from _pytest.tmpdir import TempdirFactory
+
+CaptureFixture = pytest.CaptureFixture
+MonkeyPatch = pytest.MonkeyPatch
 
 DATA_PATH: Path = Path(__file__).parent / 'data'
 OPEN_API_DATA_PATH: Path = DATA_PATH / 'openapi'
@@ -26,11 +27,9 @@ TIMESTAMP = '1985-10-26T01:21:00-07:00'
 
 @pytest.fixture(autouse=True)
 def reset_namespace(monkeypatch: MonkeyPatch):
-    monkeypatch.setattr(
-        target=__main__,
-        name='namespace',
-        value=Namespace(),
-    )
+    namespace_ = Namespace(no_color=False)
+    monkeypatch.setattr('datamodel_code_generator.__main__.namespace', namespace_)
+    monkeypatch.setattr('datamodel_code_generator.arguments.namespace', namespace_)
 
 
 @freeze_time('2019-07-26')
